requirements.txt
asgiref
Babel==2.12.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.2
dj-database-url==1.3.0
Django
django-countries==7.5.1
django-crispy-forms==2.0
django-heroku==0.3.1
django-phonenumber-field==7.0.2
django-tinymce==3.6.1
firebase-admin==6.1.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.22.0
idna==3.4
Markdown
msgpack==1.0.5
phonenumbers==8.13.10
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.22.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==1.5
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
dj-static
pymysql
mysql-connector